# perf: flex json with fast decoder