Achieved a processing time reduction of between 20% and 35% (depending on the size of the JSON input).

**BenchmarkPimoRun**

Benchmark | ns/op | B/op | allocs/op
-- | --:| --:| --:
Before | 35003 ns/op | 17382 B/op | 180 allocs/op
After | 27521 ns/op | 16176 B/op | 135 allocs/op

**BenchmarkPimoRunLarge**

Benchmark | ns/op | B/op | allocs/op
-- | --:| --:| --:
Before | 2264027 ns/op | 683047 B/op | 5905 allocs/op
After | 1494579 ns/op | 531007 B/op | 1754 allocs/op